+++ .env.production
@@ -1,0 +1,1 @@
+NEXT_PUBLIC_MARVEL_PUBLIC_KEY=175b5c70b9565986c27eb4497505acd5
