+++ components/charactersList/Skeleton.tsx
