+++ components/charactersList/index.ts
@@ -1,0 +1,2 @@
+export * from './CharactersList'
+export * from './Skeleton'
