+++ components/index.ts
@@ -1,0 +1,2 @@
+export * from './character/'
+export * from './charactersList/'
