--- pages/CharactersList/index.ts
@@ -1,4 +1,0 @@
-export {default} from './_CharactersList'
-
-
-
