+++ pages/charactersList.ts
@@ -1,2 +1,4 @@
 export { CharactersList as default } from "components/index";
 
+
+
